Rename summary -> summary-page
