+++ src/renderer/components/SummaryPage.vue
